Moved the helper methods to a new static TokenUtils class
